lib: convert prefixlen to 16-bit integer

+++ zebra/connected.c
@@ -275,7 +275,7 @@ void connected_up(struct interface *ifp, struct connected *ifc)
 
 /* Add connected IPv4 route to the interface. */
 void connected_add_ipv4(struct interface *ifp, int flags, struct in_addr *addr,
-			uint8_t prefixlen, struct in_addr *broad,
+			uint16_t prefixlen, struct in_addr *broad,
 			const char *label)
 {
 	struct prefix_ipv4 *p;
@@ -473,7 +473,7 @@ static void connected_delete_helper(struct connected *ifc, struct prefix *p)
 
 /* Delete connected IPv4 route to the interface. */
 void connected_delete_ipv4(struct interface *ifp, int flags,
-			   struct in_addr *addr, uint8_t prefixlen,
+			   struct in_addr *addr, uint16_t prefixlen,
 			   struct in_addr *broad)
 {
 	struct prefix p, d;
@@ -499,7 +499,7 @@ void connected_delete_ipv4(struct interface *ifp, int flags,
 
 /* Add connected IPv6 route to the interface. */
 void connected_add_ipv6(struct interface *ifp, int flags, struct in6_addr *addr,
-			struct in6_addr *broad, uint8_t prefixlen,
+			struct in6_addr *broad, uint16_t prefixlen,
 			const char *label)
 {
 	struct prefix_ipv6 *p;
@@ -556,7 +556,7 @@ void connected_add_ipv6(struct interface *ifp, int flags, struct in6_addr *addr,
 }
 
 void connected_delete_ipv6(struct interface *ifp, struct in6_addr *address,
-			   struct in6_addr *broad, uint8_t prefixlen)
+			   struct in6_addr *broad, uint16_t prefixlen)
 {
 	struct prefix p, d;
 	struct connected *ifc;
